fix(docker): read entire image pull response to guarantee image download
If an image did not exist on the system, the docker image command would
not finish before all the test commands would run.  This lead to various
nil pointer exceptions in the tests.

Once the entire response from the image pull is read, then the container
is now on the system.