Implemented function to create-only container
We need the ability to create a container without running it.
This PR introduced this feature with the method `CreateContainer`.
It works like the `RunContainer` but without executing the container
cmd.

You can run it as you wish with the function `container.Run(ctx)`.

The `Run(ctx)` function doesn't need a wait strategy. If you need one
you need to set it up by yourself programmatically.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>