Use docker daemon host to determine container host
This nearly represents what the Java implementation is doing.
(They currently have extra handling if inside a container.)

It should be sufficient for most use cases because when talking
to a Docker daemon this is most likely a single-host setup.
Since we are not doing Swarm or Kubernetes with this setup, multi-
host setups are not important.