feature: Context as io.Reader
resolves #110

Added a ContextArchive attribute to the FromDockerfile struct, this
allows a user to pass in an io.Reader as the Docker build context.
A user will now be allowed to create a dynamic build in code and send it
to the build.

Added test cases for both scenarios (using a local filepath and an
archive).  Added example to README.