feat: implement skip utility function for testing
I implement a new function to use as testing utility.
It skips the test if the provider is not reachable or well provisioned.

You have to declare it in your test, so if you depend on Docker and you
know you need it just keep doing as you are doing today.

In my use case we use testcontainers only for some particular tests, and
I do not like the idea to mark the testsuite as `FAILED` if a colleague
does not have Docker enabled in their local environment. Worst case if
something has to fail CI/CD will tell the truth.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>