fix: use nginx other that localstack/localstack
master is red and the cause looks like a flaky test introduced by #198.
I can't point the flakyness to code itself. And there are not race
detected running `go test --race`.

I spot that we are using a different image
`localstack/localstack:latest` for that test. it is also very big 500mb.

I changed it with nginx to keep consistency.