Add waitForSqlWithHost implementation.
This implementation mostly copies the prior waitForSql implementation,
the difference being that the waitForSqlWithHost allows the user to
construct a connection string using the host of the DB container. This
is especially useful in a scenario like CI (where our tests are being
executed in a container) or even remote Docker hosts.

To preserve backwards-compatibility, and because of the way that
`waitForSql` works, `waitForSqlWithHost` replicates the existing
behaviour mostly. In a v1 of testcontainers-go, you would probably want
to unify these two differing implementations.