Swap incorrect key and value of BindMounts and VolumeMounts
It is currently not possible to mount a local path to multiple paths in
the container.

BindMounts are currently stored in a `map[string]string` with the
hostPath being the key and containerPath as the value.
This prevents me from mounting a hostPath to multiple paths in the
container. Using a map actually makes sense because it is not possible
to mount multiple host paths to the same path in the container. Key and
value are just swapped. Same with VolumeMounts.

Yes, i know, this a huge breaking change and probably not the best fix.
We could also add an additional map like `BindMountsFixed` and deprecate
the old one. But as this library is still in version 0, a breaking
change might not be a problem?