chore: support for titles in examples (#775)
* chore: support for titles in examples

It will allow creating an example but overriding the default name with a
custom title that follows the same rules that a name, i.e. alphabetical
characters, only. The only difference is that it allows uppercase chars
that will be used when certain initials are needed in uppercase: i.e.
mongodb and MongoDB (before this changes it was Mongodb).

This title will be used here:
- the CI descriptor, as its name
- the H1 heading in the docs page for the example
- the struct for the example, which is now exampleDBContainer
- the setupExample function name, which is now setupExampleDB

* chore: wording

Co-authored-by: Kevin Wittek <kiview@users.noreply.github.com>

Co-authored-by: Kevin Wittek <kiview@users.noreply.github.com>