fix: synchronize consumer in tests
The consumer's Accept method is running in a different gorutine than
the main test.
We need to synchronize access to the messages slice.

Some tests waited for Ack before accessing Msgs and were properly
synchronized.

TestContainerLogWithErrClosed accessed Msgs while the consumer was
running and the memory access was not synchronized.

Test_StartStop did not wait for the expected lines to appear before
stopping the log producer, there was nothing to guarantee that
at least one line would be read.
The test passed because currently the producer does not interrupt
outstanding HTTP requests to fetch logs, so it implicitly waited for
some logs to be received.
We now wait for the mlem line to appear before stopping the log
producer.