fix: fix k3s module when running from inside a Docker container
The k3s module previously didn't work properly inside of a Docker
container; the caller would connect to its HTTPS server via the Docker
network gateway, which wasn't included in the list of SANS from the TLS
certificate exposed by the k3s container.

This change determines what the DaemonHost address is prior to creating
the container request, and then injects that address as one of the TLS
SANs.

When not running in a Docker container, the k3s module continues to
function as it did before, as the DaemonHost address in that scenario is
"localhost".