Robust wait for by using SQL query
Cockroach wasn't being consistent with log messages to use that as a
reliable mechanism across the different authentication methods.

This commit makes an actual SQL query to validate CRDB is correctly
running.

So we can make use of sql.DB driver and `wait.ForSQL` we've had to
register a connection URL that includes TLS config, as there's no way to
do that through standard connection strings.