fix: Add HTTPStrategy WithForcedIPv4LocalHost To Fix Docker Port Map (#1775)
* Add HTTPStrategy WithForcedIPv4LocalHost To Fix Docker Port Map Bug
Associated docker port mapping bugs:
https://github.com/moby/moby/issues/42442
https://github.com/moby/moby/issues/42375
If ipv6 is enabled in docker then these bugs affect this library. This is even if we build the docker network with ipv6 disabled since the ipv6 ports are still forwarded. This creates the potential for localhost for a container to be mapped to two different ports between ipv4 and ipv6. This is fine if you only have one container but once you have multiple containers spun up these ports can overlap where one containers ipv4 port is the same as another containers ipv6 port, at which point if you use localhost you are not guaranteed the ipv4 address and thus can end up calling into the wrong container.

* chore: rename variable

* chore: add print for the assertion

---------

Co-authored-by: Manuel de la Peña <mdelapenya@gmail.com>