chore(docker_mounts): stop doing misleading logging
As things are, the logger statement here always triggers for normal use cases.
For example, if I start a container with this mount:
testcontainers.VolumeMount("some-vol", "/some/dir")

then I get this weird message in my logs
Mount type %!s(testcontainers.MountType=1) is not supported by Testcontainers for Go

because a GenericVolumeMountSource doesn't have an implementation of
`GetVolumeOptions() *mount.VolumeOptions`

I suggest removing that logging, as in this PR, because it's going to be purely
confusing for 99% of users.