Introduce a `testcontainers_test` package to solve some of the docs imports
As discussed in #2180, the imported examples from tests for the docs from lack the package prefix before the imports, which leads to the example not being directly copy/paste-able. To solve the issue it was suggested to create a `testcontainers_test` package for the example tests and to convert everything.

I did my best to convert all the tests in a way that makes sense, and I extracted any test that would import very internal things that we should not export. Some of the tests had to be slightly rewritten (the lifecycle tests in particular), but they should stay pretty much the same.

Test by running all the root unit tests, they should all work as expected. The examples in the docs should all have the packages as prefixes now.