docs(postgres): recommend ForSQL strategy
While trying to use a postgres containers and using ForLog(), we had a
tests that worked locally, but failed on the CI, for some reason.
Searching the web, one can stumble upon https://www.xavierbouclet.com/2022/08/08/TestContainer-Wait-DB-SQL.html
If somebody went through the trouble of writing a whole blog post about
this, chances are this issue is not uncommon.
Also, the SQL strategy was contributed after the log strategy, which
might explain why it wasn't default recommendation.