feat: authenticate docker on PullImage
I'm using k3s and specifically its LoadImages function to avoid hassle
for images not published to a public repository. Through this, I noticed
we were accidentally only authenticating docker pulls when launching a
container.

This authenticates the same way, regardless of if you are launching a
container or pulling its image. This also improves the LoadImages test
as before, the pod could have not started due to an image problem, yet
still pass the test.

Signed-off-by: Adrian Cole <adrian@tetrate.io>