docs: add documentation for Exec method
Adds documentation for the Exec method in the DockerContainer struct.
This method executes a command in the current container and
returns the exit status of the executed command, an io.Reader containing
the combined stdout and stderr, and any encountered error.
The documentation emphasizes that reading directly from the io.Reader
may result in unexpected bytes due to custom stream multiplexing headers.
It recommends using stdcopy.StdCopy() from the Docker API to properly
separate the stdout and stderr.

See: https://gitlab.com/gitlab-org/gitlab-runner/-/merge_requests/503