Fix issues in BuildImage() (#2626)
* Don't accumulate errors

The code already logs them so no need to return them all to the caller - the calling code will likely log them too. This will result in the same errors being logged twice. This is confusing for the user.

* RetryNotify -> RetryNotifyWithData

* Do not reuse buildOptions.Context

buildOptions.Context cannot be reused because it might have been partially consumed by the failed ImageBuild(). So, the next invocation will send a partial context, which will not work correctly.

* Always close the response body

* Always close the build context

This ensures all resources are released properly.