fix: remove configuration reading from logger initialization
Due to the addition of a new init function for the logger initialization
and Ryuk disable message, the environment variables were reading as an
init function side effect. This caused variables set on the code using
the `os` package does not read property.

It removes the reading for the environment variables from the logger
initialization, and prints the message in a more appropriate location.