Redpanda Module: Add option for arbitrary bootstrap config
WithBootstrapConfig adds an arbitrary config kvp to the Redpanda container.
Per the name, this config will be interpolated into the generated bootstrap
config file, which is particularly useful for configs requiring a restart
when otherwise applied to a running Redpanda instance.

Includes a test that sets a couple of restart-requiring cluster configs and
checks a) that the values are reflected in the config state at run time AND
b) that the cluster does not need a restart. (b) demonstrates that the configs
were bootstrapped rather than applied to a live cluster.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>