debugging: for termination issue
Add debug logging to help identify issue

More debug to see if reaper is cause the issue.

Disable ryuk to see if that fixes the tests.

Add docker version info to see if there's a bug in the action runners
docker version.

Add docker info, config to see if we can enable debug on docker daemon.

Restart docker with debug so we can get request logs.