feat: merge kafka native into kafka module
This should make it easier for users to start.
The change attempts to steer new users to
start with native image, which shows significant
performance improvement especially if many
containers need to be started during entire
test, due to GraalVM optimizations.
While new users should prefer using Kafka
Native, these changes are made in a non-breaking
fashion such that if image was not detected as
being from Apache, it is assumed to be from
Confluent, so that even custom images that
used to work before should continue to work,
in case if they were based on confluentinc images